Renamed Trigonometry file. Original name was misspelt.
