+++ functions/trigonometry.py
